+++ web/src/components/Reveal.tsx
@@ -1,0 +1,47 @@
+import { useEffect, useRef, useState } from "react"
+
+export default function Reveal({
+    children,
+    delay = 0,
+    className = '',
+}: {
+    children: React.ReactNode
+    delay?: number
+    className?: string
+}) {
+    const { ref, visible } = useInView(0.1)
+    return (
+        <div
+            ref={ref}
+            className={className}
+            style={{
+                opacity: visible ? 1 : 0,
+                transform: visible ? 'translateY(0)' : 'translateY(28px)',
+                transition: `opacity 0.65s cubic-bezier(0.16,1,0.3,1) ${delay}s, transform 0.65s cubic-bezier(0.16,1,0.3,1) ${delay}s`,
+            }}
+        >
+            {children}
+        </div>
+    )
+}
+
+function useInView(threshold = 0.15) {
+    const ref = useRef<HTMLDivElement>(null)
+    const [visible, setVisible] = useState(false)
+    useEffect(() => {
+        const el = ref.current
+        if (!el) return
+        const obs = new IntersectionObserver(
+            ([e]) => {
+                if (e.isIntersecting) {
+                    setVisible(true)
+                    obs.disconnect()
+                }
+            },
+            { threshold },
+        )
+        obs.observe(el)
+        return () => obs.disconnect()
+    }, [threshold])
+    return { ref, visible }
+}
